Move Cloud Run worker-ID plugin into contrib/gcp/cloudrun/workerid submodule
Move the Cloud Run worker-identity plugin out of module
contrib/gcp/cloudrun into a new contrib/gcp/cloudrun/workerid submodule
and rename its Go package from cloudrun to workerid, so the cloudrun
directory can also host the separate OpenTelemetry plugin (also package
cloudrun) without a hard Go package collision.

- git mv the module files under contrib/gcp/cloudrun/workerid/
- module path -> go.temporal.io/sdk/contrib/gcp/cloudrun/workerid; fix
  the replace directive to ../../../../ (now four levels up)
- package cloudrun -> package workerid across all .go files; update
  doc.go and README import paths and usage to workerid.NewPlugin
- drop the leading HTML comment block from CHANGELOG.md so it starts at
  # Changelog / ## [Unreleased] / ### Added

The plugin type (Plugin/NewPlugin/PluginOptions), the Metadata reader,
the plugin name constant, and the "cloudrun:" error prefixes (which
refer to the Cloud Run platform) are unchanged.
