+++ contrib/gcp/cloudrun/workerid/CHANGELOG.md
@@ -1,12 +1,3 @@
-<!--
-Release notes for go.temporal.io/sdk/contrib/gcp/cloudrun.
-Loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
-
-Add user-facing changes below under the appropriate heading (create the heading
-if it does not yet exist): Added, Changed, Deprecated, Breaking Changes, Fixed,
-or Security.
--->
-
 # Changelog
 
 ## [Unreleased]
