+++ contrib/gcp/cloudrun/workerid/metadata.go
@@ -1,4 +1,4 @@
-package cloudrun
+package workerid
 
 import (
 	"context"
