+++ contrib/gcp/cloudrun/workerid/plugin.go
@@ -1,4 +1,4 @@
-package cloudrun
+package workerid
 
 import (
 	"context"
